chore(repo): Rename `tests/utils` to `tests/test_utils` to prevent confusion.
